backend/requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
django-allauth>=0.54.0
python-dotenv>=1.0.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
protobuf>=4.21.0
PyJWT>=2.8.0
cryptography>=41.0.0
Django==4.2.7
requests==2.31.0
pandas
openpyxl
reportlab